--- react/src/components/WebsocketNotifications/index.ts
@@ -1,1 +1,0 @@
-export { default } from './WebsocketNotifications';
